+++ .vscode/settings.json
@@ -9,5 +9,8 @@
     "source.fixAll.eslint": true
     },
     "cSpell.words": [
-    ]
+    ],
+    "[javascriptreact]": {
+      "editor.defaultFormatter": "vscode.typescript-language-features"
+    }
   }
