+++ src/components/footers/forfeit-match-footer/ForfeitMatchFooter.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import ForfeitMatchButton from '../../buttons/forfeit-match-button/ForfeitMatchButton'
+
+function ForfeitMatchFooter () {
+  return (
+    <footer className='absolute bottom-4 right-4'>
+      <ForfeitMatchButton />
+    </footer>
+  )
+}
+
+export default ForfeitMatchFooter
