+++ src/components/footers/home-footer/HomeFooter.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import FooterButton from '../../buttons/footer-button/FooterButton'
+
+function HomeFooter () {
+  return (
+    <footer className='absolute bottom-4 left-4'>
+      <FooterButton text={'Report Bug'} path={'/'} />
+    </footer>
+  )
+}
+
+export default HomeFooter
